+++ tests/appsec/api_security/test_schemas.py
@@ -36,13 +36,12 @@ def equal_value(t1, t2):
 
 @rfc("https://docs.google.com/document/d/1OCHPBCAErOL2FhLl64YAHB8woDyq66y5t-JGolxdf1Q/edit#heading=h.bth088vsbjrz")
 @released(
-    cpp="?",
     dotnet="?",
     golang="?",
     java="?",
     nodejs="?",
     php_appsec="?",
-    python={"django-poc": "1.16", "flask-poc": "1.16", "*": "1.16.1"},
+    python={"django-poc": "1.18", "flask-poc": "1.18", "*": "?"},
     ruby="?",
 )
 @coverage.basic
@@ -64,13 +63,12 @@ class Test_Schema_Request_Headers:
 
 @rfc("https://docs.google.com/document/d/1OCHPBCAErOL2FhLl64YAHB8woDyq66y5t-JGolxdf1Q/edit#heading=h.bth088vsbjrz")
 @released(
-    cpp="?",
     dotnet="?",
     golang="?",
     java="?",
     nodejs="?",
     php_appsec="?",
-    python={"django-poc": "1.16", "flask-poc": "1.16", "*": "1.16.1"},
+    python={"django-poc": "1.18", "flask-poc": "1.18", "*": "?"},
     ruby="?",
 )
 @coverage.basic
@@ -94,13 +92,12 @@ class Test_Schema_Request_Query_Parameters:
 
 @rfc("https://docs.google.com/document/d/1OCHPBCAErOL2FhLl64YAHB8woDyq66y5t-JGolxdf1Q/edit#heading=h.bth088vsbjrz")
 @released(
-    cpp="?",
     dotnet="?",
     golang="?",
     java="?",
     nodejs="?",
     php_appsec="?",
-    python={"django-poc": "1.16", "flask-poc": "1.16", "*": "1.16.1"},
+    python={"django-poc": "1.18", "flask-poc": "1.18", "*": "?"},
     ruby="?",
 )
 @coverage.basic
@@ -126,13 +123,12 @@ class Test_Schema_Request_Path_Parameters:
 
 @rfc("https://docs.google.com/document/d/1OCHPBCAErOL2FhLl64YAHB8woDyq66y5t-JGolxdf1Q/edit#heading=h.bth088vsbjrz")
 @released(
-    cpp="?",
     dotnet="?",
     golang="?",
     java="?",
     nodejs="?",
     php_appsec="?",
-    python={"django-poc": "1.16", "flask-poc": "1.16", "*": "1.16.1"},
+    python={"django-poc": "1.18", "flask-poc": "1.18", "*": "?"},
     ruby="?",
 )
 @coverage.basic
@@ -153,13 +149,12 @@ class Test_Schema_Request_Body:
 
 @rfc("https://docs.google.com/document/d/1OCHPBCAErOL2FhLl64YAHB8woDyq66y5t-JGolxdf1Q/edit#heading=h.bth088vsbjrz")
 @released(
-    cpp="?",
     dotnet="?",
     golang="?",
     java="?",
     nodejs="?",
     php_appsec="?",
-    python={"django-poc": "1.16", "flask-poc": "1.16", "*": "1.16.1"},
+    python={"django-poc": "1.18", "flask-poc": "1.18", "*": "?"},
     ruby="?",
 )
 @coverage.basic
@@ -183,7 +178,7 @@ class Test_Schema_Reponse_Headers:
 
 @rfc("https://docs.google.com/document/d/1OCHPBCAErOL2FhLl64YAHB8woDyq66y5t-JGolxdf1Q/edit#heading=h.bth088vsbjrz")
 @released(
-    cpp="?", dotnet="?", golang="?", java="?", nodejs="?", php_appsec="?", python="?", ruby="?",
+    dotnet="?", golang="?", java="?", nodejs="?", php_appsec="?", python="?", ruby="?",
 )
 @coverage.not_implemented
 @scenarios.appsec_api_security