+++ tests/parametric/test_headers_b3.py
@@ -12,21 +12,39 @@ parametrize = pytest.mark.parametrize
 
 
 def enable_b3() -> Any:
-    env1 = {
-        "DD_TRACE_PROPAGATION_STYLE_EXTRACT": "B3 SINGLE HEADER",
-        "DD_TRACE_PROPAGATION_STYLE_INJECT": "b3 single header",
+    env = {
+        "DD_TRACE_PROPAGATION_STYLE_EXTRACT": "B3 single header",
+        "DD_TRACE_PROPAGATION_STYLE_INJECT": "B3 single header",
     }
-    env2 = {
+    return parametrize("library_env", [env])
+
+
+def enable_b3_single_key() -> Any:
+    env = {
         "DD_TRACE_PROPAGATION_STYLE": "B3 single header",
     }
-    return parametrize("library_env", [env1, env2])
+    return parametrize("library_env", [env])
+
+
+def enable_migrated_b3() -> Any:
+    env = {
+        "DD_TRACE_PROPAGATION_STYLE_EXTRACT": "b3",
+        "DD_TRACE_PROPAGATION_STYLE_INJECT": "b3",
+    }
+    return parametrize("library_env", [env])
+
+
+def enable_migrated_b3_single_key() -> Any:
+    env = {
+        "DD_TRACE_PROPAGATION_STYLE": "b3",
+    }
+    return parametrize("library_env", [env])
 
 
 @scenarios.parametric
 class Test_Headers_B3:
     @enable_b3()
     @missing_feature(context.library == "cpp", reason="format of DD_TRACE_PROPAGATION_STYLE_EXTRACT not supported")
-    @missing_feature(context.library == "ruby", reason="Ruby doesn't support case-insensitive distributed headers")
     def test_headers_b3_extract_valid(self, test_agent, test_library):
         """Ensure that b3 distributed tracing headers are extracted
         and activated properly.
@@ -57,7 +75,6 @@ class Test_Headers_B3:
 
     @enable_b3()
     @missing_feature(context.library == "cpp", reason="format of DD_TRACE_PROPAGATION_STYLE_EXTRACT not supported")
-    @missing_feature(context.library == "ruby", reason="Ruby doesn't support case-insensitive distributed headers")
     def test_headers_b3_inject_valid(self, test_agent, test_library):
         """Ensure that b3 distributed tracing headers are injected properly.
         """
@@ -78,8 +95,7 @@ class Test_Headers_B3:
 
     @enable_b3()
     @missing_feature(context.library == "cpp", reason="format of DD_TRACE_PROPAGATION_STYLE_EXTRACT not supported")
-    @missing_feature(context.library == "ruby", reason="Ruby doesn't support case-insensitive distributed headers")
-    def test_headers_b3multi_propagate_valid(self, test_agent, test_library):
+    def test_headers_b3_propagate_valid(self, test_agent, test_library):
         """Ensure that b3 distributed tracing headers are extracted
         and injected properly.
         """
@@ -102,8 +118,7 @@ class Test_Headers_B3:
 
     @enable_b3()
     @missing_feature(context.library == "cpp", reason="format of DD_TRACE_PROPAGATION_STYLE_EXTRACT not supported")
-    @missing_feature(context.library == "ruby", reason="Ruby doesn't support case-insensitive distributed headers")
-    def test_headers_b3multi_propagate_invalid(self, test_agent, test_library):
+    def test_headers_b3_propagate_invalid(self, test_agent, test_library):
         """Ensure that invalid b3 distributed tracing headers are not extracted
         and the new span context is injected properly.
         """
@@ -124,3 +139,86 @@ class Test_Headers_B3:
         assert int(b3_span_id, base=16) == span.get("span_id") and len(b3_span_id) == 16
         assert b3_sampling == "1" if span["metrics"].get(SAMPLING_PRIORITY_KEY) > 0 else "0"
         assert span["meta"].get(ORIGIN) is None
+
+    @enable_b3_single_key()
+    @missing_feature(context.library == "cpp", reason="format of DD_TRACE_PROPAGATION_STYLE_EXTRACT not supported")
+    @missing_feature(
+        context.library == "ruby", reason="Propagators not configured for DD_TRACE_PROPAGATION_STYLE config",
+    )
+    def test_headers_b3_single_key_propagate_valid(self, test_agent, test_library):
+        self.test_headers_b3_propagate_valid(test_agent, test_library)
+
+    @enable_migrated_b3()
+    @missing_feature(context.library == "cpp", reason="format of DD_TRACE_PROPAGATION_STYLE_EXTRACT not supported")
+    @missing_feature(context.library == "dotnet", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "golang", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "java", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "nodejs", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "php", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "python", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "python_http", reason="Need to remove b3=b3multi alias")
+    def test_headers_b3_migrated_extract_valid(self, test_agent, test_library):
+        self.test_headers_b3_extract_invalid(test_agent, test_library)
+
+    @enable_migrated_b3()
+    @missing_feature(context.library == "cpp", reason="format of DD_TRACE_PROPAGATION_STYLE_EXTRACT not supported")
+    @missing_feature(context.library == "dotnet", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "golang", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "java", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "nodejs", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "php", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "python", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "python_http", reason="Need to remove b3=b3multi alias")
+    def test_headers_b3_migrated_extract_invalid(self, test_agent, test_library):
+        self.test_headers_b3_extract_invalid(test_agent, test_library)
+
+    @enable_migrated_b3()
+    @missing_feature(context.library == "cpp", reason="format of DD_TRACE_PROPAGATION_STYLE_EXTRACT not supported")
+    @missing_feature(context.library == "dotnet", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "golang", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "java", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "nodejs", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "php", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "python", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "python_http", reason="Need to remove b3=b3multi alias")
+    def test_headers_b3_migrated_inject_valid(self, test_agent, test_library):
+        self.test_headers_b3_inject_valid(test_agent, test_library)
+
+    @enable_migrated_b3()
+    @missing_feature(context.library == "cpp", reason="format of DD_TRACE_PROPAGATION_STYLE_EXTRACT not supported")
+    @missing_feature(context.library == "dotnet", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "golang", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "java", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "nodejs", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "php", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "python", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "python_http", reason="Need to remove b3=b3multi alias")
+    def test_headers_b3_migrated_propagate_valid(self, test_agent, test_library):
+        self.test_headers_b3_propagate_valid(test_agent, test_library)
+
+    @enable_migrated_b3()
+    @missing_feature(context.library == "cpp", reason="format of DD_TRACE_PROPAGATION_STYLE_EXTRACT not supported")
+    @missing_feature(context.library == "dotnet", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "golang", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "java", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "nodejs", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "php", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "python", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "python_http", reason="Need to remove b3=b3multi alias")
+    def test_headers_b3_migrated_propagate_invalid(self, test_agent, test_library):
+        self.test_headers_b3_propagate_invalid(test_agent, test_library)
+
+    @enable_migrated_b3_single_key()
+    @missing_feature(context.library == "cpp", reason="format of DD_TRACE_PROPAGATION_STYLE_EXTRACT not supported")
+    @missing_feature(context.library == "dotnet", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "golang", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "java", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "nodejs", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "php", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "python", reason="Need to remove b3=b3multi alias")
+    @missing_feature(context.library == "python_http", reason="Need to remove b3=b3multi alias")
+    @missing_feature(
+        context.library == "ruby", reason="Propagators not configured for DD_TRACE_PROPAGATION_STYLE config",
+    )
+    def test_headers_b3_migrated_single_key_propagate_valid(self, test_agent, test_library):
+        self.test_headers_b3_propagate_valid(test_agent, test_library)