+++ utils/_context/containers.py
@@ -14,6 +14,7 @@ import requests
 
 from utils._context.library_version import LibraryVersion, Version
 from utils.tools import logger
+from utils import interfaces
 
 
 @lru_cache
@@ -44,6 +45,7 @@ class TestedContainer:
         environment=None,
         allow_old_container=False,
         healthcheck=None,
+        stdout_interface=None,
         **kwargs,
     ) -> None:
         self.name = name
@@ -56,6 +58,7 @@ class TestedContainer:
         self.environment = environment
         self.kwargs = kwargs
         self._container = None
+        self.stdout_interface = stdout_interface
 
     def configure(self, replay):
 
@@ -69,6 +72,9 @@ class TestedContainer:
             self.image.save_image_info(self.log_folder_path)
         else:
             self.image.load_from_logs(self.log_folder_path)
+
+        if self.stdout_interface is not None:
+            self.stdout_interface.configure(replay)
 
     @property
     def container_name(self):
@@ -178,32 +184,31 @@ class TestedContainer:
 
         self.kwargs["volumes"] = result
 
-    def save_logs(self):
-        if not self._container:
-            return
-
-        with open(f"{self.log_folder_path}/stdout.log", "wb") as f:
-            f.write(self._container.logs(stdout=True, stderr=False))
-
-        with open(f"{self.log_folder_path}/stderr.log", "wb") as f:
-            f.write(self._container.logs(stdout=False, stderr=True))
-
     def stop(self):
         self._container.stop()
 
     def remove(self):
+        logger.debug(f"Removing container {self.name}")
 
-        if not self._container:
-            return
+        if self._container:
+            try:
+                # collect logs before removing
+                with open(f"{self.log_folder_path}/stdout.log", "wb") as f:
+                    f.write(self._container.logs(stdout=True, stderr=False))
 
-        try:
-            self._container.remove(force=True)
-        except:
-            # Sometimes, the container does not exists.
-            # We can safely ignore this, because if it's another issue
-            # it will be killed at startup
+                with open(f"{self.log_folder_path}/stderr.log", "wb") as f:
+                    f.write(self._container.logs(stdout=False, stderr=True))
 
-            pass
+                self._container.remove(force=True)
+            except:
+                # Sometimes, the container does not exists.
+                # We can safely ignore this, because if it's another issue
+                # it will be killed at startup
+
+                pass
+
+        if self.stdout_interface is not None:
+            self.stdout_interface.load_data()
 
 
 class ImageInfo:
@@ -283,8 +288,12 @@ class AgentContainer(TestedContainer):
             name="agent",
             host_log_folder=host_log_folder,
             environment=environment,
-            healthcheck={"test": f"curl --fail http://localhost:{self.agent_port}/info", "retries": 60},
+            healthcheck={
+                "test": f"curl --fail --silent --show-error http://localhost:{self.agent_port}/info",
+                "retries": 60,
+            },
             ports={self.agent_port: f"{self.agent_port}/tcp"},
+            stdout_interface=interfaces.agent_stdout,
         )
 
         self.agent_version = None
@@ -334,8 +343,9 @@ class WeblogContainer(TestedContainer):
             # ddprof's perf event open is blocked by default by docker's seccomp profile
             # This is worse than the line above though prevents mmap bugs locally
             security_opt=["seccomp=unconfined"],
-            healthcheck={"test": f"curl --fail localhost:{weblog.port}", "retries": 60},
+            healthcheck={"test": f"curl --fail --silent --show-error localhost:{weblog.port}", "retries": 60},
             ports={"7777/tcp": weblog.port, "7778/tcp": weblog._grpc_port},
+            stdout_interface=interfaces.library_stdout,
         )
 
         self.tracer_sampling_rate = tracer_sampling_rate
@@ -425,6 +435,7 @@ class PostgresContainer(TestedContainer):
                     "mode": "ro",
                 }
             },
+            stdout_interface=interfaces.postgres,
         )
 
 
@@ -528,7 +539,9 @@ class OpenTelemetryCollectorContainer(TestedContainer):
         self._otel_config_host_path = "./utils/build/docker/otelcol-config.yaml"
 
         if "DOCKER_HOST" in os.environ:
-            self._otel_host = re.sub(r"^ssh://([^@]+@|)", "", os.environ["DOCKER_HOST"])
+            m = re.match(r"(?:ssh:|tcp:|fd:|)//(?:[^@]+@|)([^:]+)", os.environ["DOCKER_HOST"])
+            if m is not None:
+                self._otel_host = m.group(1)
         else:
             self._otel_host = "localhost"